add generateStaticParam to blockexplorer [txHash] and [address]
