+++ packages/nextjs/next.config.js
@@ -14,6 +14,7 @@ const nextConfig = {
     config.externals.push("pino-pretty", "lokijs", "encoding");
     return config;
   },
+  output: "export",
 };
 
 module.exports = nextConfig;
